+++ GCore/os/src/hal/arch/riscv/config.rs
@@ -16,7 +16,6 @@ pub const KERNEL_HEAP_SIZE: usize = PAGE_SIZE * 0x2000;
 pub const KERNEL_HEAP_SIZE: usize = PAGE_SIZE * 0x2000;
 pub const MMAP_BASE: usize = 0x6000_0000;
 pub const MMAP_END: usize = 0x8000_0000;
-pub const VDSO_BASE: usize = 0x3000_0000;
 pub const SKIP_NUM: usize = 2;
 
 // manually make usable memory space equal